METRON-2164 Remove the Split-Join Enrichment Topology
