+++ metron-deployment/packaging/ambari/.gitignore
@@ -4,6 +4,4 @@ elasticsearch.properties.j2
 solr.properties.j2
 hdfs.properties.j2
 enrichment.properties.j2
-enrichment-splitjoin.properties.j2
-enrichment-unified.properties.j2
 pcap.properties.j2
