+++ metron-deployment/packaging/ambari/metron-mpack/pom.xml
@@ -110,8 +110,7 @@
                                 <resource>
                                     <directory>${basedir}/../../../../metron-platform/metron-enrichment/metron-enrichment-storm/src/main/config</directory>
                                     <includes>
-                                        <include>enrichment-splitjoin.properties.j2</include>
-                                        <include>enrichment-unified.properties.j2</include>
+                                        <include>enrichment.properties.j2</include>
                                     </includes>
                                     <filtering>false</filtering>
                                 </resource>
@@ -185,8 +184,7 @@
                         <fileset>
                             <directory>${basedir}/src/main/resources/common-services/METRON/CURRENT/package/templates</directory>
                             <includes>
-                                <include>enrichment-unified.properties.j2</include>
-                                <include>enrichment-splitjoin.properties.j2</include>
+                                <include>enrichment.properties.j2</include>
                                 <include>elasticsearch.properties.j2</include>
                                 <include>hdfs.properties.j2</include>
                             </includes>
