+++ metron-deployment/packaging/ambari/metron-mpack/src/main/resources/common-services/METRON/CURRENT/configuration/metron-enrichment-env.xml
@@ -193,88 +193,6 @@
       <empty-value-valid>true</empty-value-valid>
     </value-attributes>
   </property>
-  <property>
-    <name>enrichment_topology</name>
-    <description>Which Enrichment topology to execute. Note: Split-Join is deprecated in favor of the Unified topology.</description>
-    <value>Unified</value>
-    <display-name>Enrichment Topology</display-name>
-    <value-attributes>
-      <type>value-list</type>
-      <entries>
-        <entry>
-          <value>Unified</value>
-        </entry>
-        <entry>
-          <value>Split-Join</value>
-        </entry>
-      </entries>
-      <selection-cardinality>1</selection-cardinality>
-    </value-attributes>
-  </property>
-
-  <!--
-    split-join topology parameters
-  -->
-  <property>
-    <name>enrichment_join_cache_size</name>
-    <description>Enrichment join bolt max cache size for the Split Join Enrichment Topology</description>
-    <value>100000</value>
-    <display-name>Enrichment Join Max Cache Size</display-name>
-  </property>
-  <property>
-    <name>threatintel_join_cache_size</name>
-    <description>Threat Intel join bolt max cache size for the Split Join Enrichment Topology</description>
-    <value>100000</value>
-    <display-name>Threat Intel Join Max Cache Size</display-name>
-  </property>
-  <property>
-    <name>enrichment_kafka_spout_parallelism</name>
-    <description>Kafka spout parallelism for the Split Join Enrichment Topology</description>
-    <value>1</value>
-    <display-name>Enrichment Spout Parallelism</display-name>
-  </property>
-  <property>
-    <name>enrichment_split_parallelism</name>
-    <description>Enrichment split bolt parallelism for the Split Join Enrichment Topology</description>
-    <value>1</value>
-    <display-name>Enrichment Split Parallelism</display-name>
-  </property>
-  <property>
-    <name>enrichment_stellar_parallelism</name>
-    <description>Stellar enrichment bolt parallelism for the Split Join Enrichment Topology</description>
-    <value>1</value>
-    <display-name>Stellar Enrichment Parallelism</display-name>
-  </property>
-  <property>
-    <name>enrichment_join_parallelism</name>
-    <description>Enrichment join bolt parallelism for the Split Join Enrichment Topology</description>
-    <value>1</value>
-    <display-name>Enrichment Join Parallelism</display-name>
-  </property>
-  <property>
-    <name>threat_intel_split_parallelism</name>
-    <description>Threat Intel split bolt parallelism for the Split Join Enrichment Topology</description>
-    <value>1</value>
-    <display-name>Threat Intel Split Parallelism</display-name>
-  </property>
-  <property>
-    <name>threat_intel_stellar_parallelism</name>
-    <description>Threat Intel stellar bolt parallelism for the Split Join Enrichment Topology</description>
-    <value>1</value>
-    <display-name>Threat Intel Stellar Parallelism</display-name>
-  </property>
-  <property>
-    <name>threat_intel_join_parallelism</name>
-    <description>Threat Intel join bolt parallelism for the Split Join Enrichment Topology</description>
-    <value>1</value>
-    <display-name>Threat Intel Join Parallelism</display-name>
-  </property>
-  <property>
-    <name>kafka_writer_parallelism</name>
-    <description>Kafka writer parallelism for the Split Join Enrichment Topology</description>
-    <value>1</value>
-    <display-name>Enrichment Kafka Writer Parallelism</display-name>
-  </property>
 
   <!--
     unified topology parameters
