+++ metron-deployment/packaging/ambari/metron-mpack/src/main/resources/common-services/METRON/CURRENT/themes/metron_theme.json
@@ -65,7 +65,7 @@
             "display-name": "Enrichment",
             "layout": {
               "tab-columns": "1",
-              "tab-rows": "5",
+              "tab-rows": "4",
               "sections": [
                 {
                   "name": "section-enrichment-adapters",
@@ -136,25 +136,6 @@
                     {
                       "name": "subsection-enrichment-unified",
                       "display-name": "Unified Topology",
-                      "row-index": "0",
-                      "column-index": "0",
-                      "row-span": "1",
-                      "column-span": "1"
-                    }
-                  ]
-                },
-                {
-                  "name": "section-enrichment-splitjoin",
-                  "row-index": "4",
-                  "column-index": "0",
-                  "row-span": "1",
-                  "column-span": "1",
-                  "section-columns": "1",
-                  "section-rows": "1",
-                  "subsections": [
-                    {
-                      "name": "subsection-enrichment-splitjoin",
-                      "display-name": "Split Join Topology",
                       "row-index": "0",
                       "column-index": "0",
                       "row-span": "1",
@@ -609,46 +590,6 @@
         {
           "config": "metron-enrichment-env/enrichment_topology",
           "subsection-name": "subsection-enrichment-storm"
-        },
-        {
-          "config": "metron-enrichment-env/enrichment_join_cache_size",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/threatintel_join_cache_size",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/enrichment_kafka_spout_parallelism",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/enrichment_split_parallelism",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/enrichment_stellar_parallelism",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/enrichment_join_parallelism",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/threat_intel_split_parallelism",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/threat_intel_stellar_parallelism",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/threat_intel_join_parallelism",
-          "subsection-name": "subsection-enrichment-splitjoin"
-        },
-        {
-          "config": "metron-enrichment-env/kafka_writer_parallelism",
-          "subsection-name": "subsection-enrichment-splitjoin"
         },
         {
           "config": "metron-enrichment-env/unified_kafka_spout_parallelism",
